// packages/design-vue/src/components/molecules/taskin-arm-with-phone/TaskinArmWithPhone.types.ts
export interface TaskinArmWithPhoneProps {
  armColor?: string;
  phoneColor?: string;
  screenColor?: string;
  phoneWidth?: number;
  phoneHeight?: number;
  phoneOnLeft?: boolean;
  phoneOnRight?: boolean;
  phoneOffsetX?: number;
  phoneOffsetY?: number;
  leftArmRotation?: number;
  rightArmRotation?: number;
  animationsEnabled?: boolean;
}